+++ app/src/androidTest/java/ru/kravchenkoanatoly/weatherApp/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package ru.kravchenkoanatoly.weatherapp
+package ru.kravchenkoanatoly.weatherApp
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
