+++ app/src/main/java/ru/kravchenkoanatoly/weatherApp/ui/gallery/GalleryFragment.kt
@@ -1,4 +1,4 @@
-package ru.kravchenkoanatoly.weatherapp.ui.home
+package ru.kravchenkoanatoly.weatherApp.ui.gallery
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,11 +7,11 @@ import android.view.ViewGroup
 import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
-import ru.kravchenkoanatoly.weatherapp.databinding.FragmentHomeBinding
+import ru.kravchenkoanatoly.weatherapp.databinding.FragmentGalleryBinding
 
-class HomeFragment : Fragment() {
+class GalleryFragment : Fragment() {
 
-    private var _binding: FragmentHomeBinding? = null
+    private var _binding: FragmentGalleryBinding? = null
 
     // This property is only valid between onCreateView and
     // onDestroyView.
@@ -22,14 +22,14 @@ class HomeFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        val homeViewModel =
-                ViewModelProvider(this).get(HomeViewModel::class.java)
+        val galleryViewModel =
+                ViewModelProvider(this).get(GalleryViewModel::class.java)
 
-        _binding = FragmentHomeBinding.inflate(inflater, container, false)
+        _binding = FragmentGalleryBinding.inflate(inflater, container, false)
         val root: View = binding.root
 
-        val textView: TextView = binding.textHome
-        homeViewModel.text.observe(viewLifecycleOwner) {
+        val textView: TextView = binding.textGallery
+        galleryViewModel.text.observe(viewLifecycleOwner) {
             textView.text = it
         }
         return root
