+++ app/src/main/java/ru/kravchenkoanatoly/weatherApp/ui/gallery/GalleryViewModel.kt
@@ -1,13 +1,13 @@
-package ru.kravchenkoanatoly.weatherapp.ui.home
+package ru.kravchenkoanatoly.weatherApp.ui.gallery
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class HomeViewModel : ViewModel() {
+class GalleryViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is home Fragment"
+        value = "This is gallery Fragment"
     }
     val text: LiveData<String> = _text
 }
