+++ app/src/main/java/ru/kravchenkoanatoly/weatherApp/ui/home/HomeViewModel.kt
@@ -1,13 +1,13 @@
-package ru.kravchenkoanatoly.weatherapp.ui.slideshow
+package ru.kravchenkoanatoly.weatherApp.ui.home
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class SlideshowViewModel : ViewModel() {
+class HomeViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is slideshow Fragment"
+        value = "This is home Fragment"
     }
     val text: LiveData<String> = _text
 }
