+++ app/src/main/java/ru/kravchenkoanatoly/weatherApp/ui/slideshow/SlideshowViewModel.kt
@@ -1,13 +1,13 @@
-package ru.kravchenkoanatoly.weatherapp.ui.gallery
+package ru.kravchenkoanatoly.weatherApp.ui.slideshow
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class GalleryViewModel : ViewModel() {
+class SlideshowViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is gallery Fragment"
+        value = "This is slideshow Fragment"
     }
     val text: LiveData<String> = _text
 }
