+++ build.gradle.kts
@@ -2,4 +2,5 @@
 plugins {
     id("com.android.application") version "8.1.2" apply false
     id("org.jetbrains.kotlin.android") version "1.9.0" apply false
+    id("org.jetbrains.kotlin.jvm") version "1.9.0" apply false
 }
