+++ settings.gradle.kts
@@ -15,4 +15,4 @@ dependencyResolutionManagement {
 
 rootProject.name = "Weather App"
 include(":app")
- 
+//include(":buildSrc")
